bigdata/requirements.txt
Django>=4.0
djangorestframework>=3.12.4
drf-yasg>=1.21.3
django-rest-swagger>=2.2.0
djangorestframework_simplejwt>=5.2.0


